Added the various starbound packages
Not sure how long the starbound packages will last. I seem to recall seeing somewhere that they wouldn't be available forever...though that may well have been some other game.

All packages (except 4-packs, which contain other packages) are listed as containing soundtracks, so they're all added that way.

I have a vague idea that it's based on the same code as terraria, so I listed it as being for PC and no other platform (despite the lack of information on the humble store).