Sales end may 15th, 26th
Both talisman items are on sale, sale ends may 26th

Super Splatters is 30% off for preorder, unknown when sale ends

Incredipede is listed as being on sale, has been since it was posted, I think, but has no end date for the sale.

Anomaly 2 preorder sale ends may 15th. The link might change.

------

I'm adding a 'sale' boolean for sales, but I'm not implementing it anywhere, it's just something that can be searched for manually, since I'm not listing the titles of sale games in the commit.