Added blade symphony
The url works both with and without the ?preview=bit, so I went ahead and added it without as I assume the preview bit will be removed later. If it doesn't work, the original URL *with* the preview text is [bladesymphony?preview=JSlsOmFvHOcQ](http://www.humblebundle.com/store/product/bladesymphony?preview=JSlsOmFvHOcQ).