Defender's Quest also comes with Desura and GOG keys
I know because I've just purchased it :)

However, it seems to me that you can redeem either the Desura key OR the GOG one, but not both. Haven't actually tried it because I don't have an account there, but the GOG key appeared as "redeemed" as soon as I activated the Desura one.